♻️Refactor: 버튼 재사용
