+++ src/pages/home/Home.jsx
@@ -1,11 +1,11 @@
-import Button from './Button';
+import Button from '../../components/Button';
 import SwiperSection from './Swiper';
 
 export default function HomePage() {
   return (
     <>
       <SwiperSection />
-      <Button />
+      <Button text="시작하기" link="/login" height={60} width={100} />
     </>
   );
 }
