+++ src/pages/home/styles.js
@@ -1,5 +1,4 @@
 import { Swiper } from 'swiper/react';
-import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 
 // Swiper.jsx
@@ -46,27 +45,5 @@ export const StyledSwiper = styled(Swiper)`
     margin-bottom: 5px;
     font-size: ${({ theme }) => theme.fontSize.s};
     font-weight: ${({ theme }) => theme.fontWeight.xs};
-  }
-`;
-
-// Button.jsx
-
-export const StyleLink = styled(Link)`
-  width: 100%;
-  height: 60px;
-  margin-top: 30px;
-`;
-
-export const StyledButton = styled.button`
-  width: 100%;
-  height: 100%;
-  border: none;
-  border-radius: 15px;
-  background-color: ${({ theme }) => theme.color.main};
-  color: white;
-
-  & p {
-    font-size: ${({ theme }) => theme.fontSize.l};
-    font-weight: ${({ theme }) => theme.fontWeight.xl};
   }
 `;
